검색어 삭제 테스트 코드 Service로 이관
